Replace c_double and c_int by double precision and integer

+++ example/fortran/dense/minij_function_mpi.f90
@@ -82,8 +82,8 @@ contains
   function minij_matrix_element(i, j) bind(c)
     use iso_c_binding
     implicit none
-    real(c_double) :: minij_matrix_element
-    integer(c_int), value, intent(in) :: i, j
+    double precision :: minij_matrix_element
+    integer, value, intent(in) :: i, j
     minij_matrix_element = dble(min(i, j))
   end function minij_matrix_element
